moved Divider to deprecated folder, and created fake Divider  pointer for backward compatibility
