+++ packages/wix-ui-core/src/components/Autocomplete/Autocomplete.st.css
@@ -4,7 +4,7 @@
 }
 
 :import {
-  -st-from: "../Divider/Divider.st.css";
+  -st-from: "../deprecated/Divider/Divider.st.css";
   -st-default: Divider;
 }
