+++ packages/wix-ui-core/src/components/Divider/index.ts
@@ -1,1 +1,2 @@
-export {Divider, DividerProps} from './Divider';
+// Backward compatibility for deprecated components
+export * from '../../components/deprecated/Divider';
