+++ packages/wix-ui-core/src/components/deprecated/Divider/Divider.driver.ts
