+++ packages/wix-ui-core/src/components/deprecated/Divider/Divider.protractor.driver.ts
@@ -1,4 +1,4 @@
-import {BaseDriver, DriverFactory} from './../../common/BaseDriver.protractor';
+import {BaseDriver, DriverFactory} from './../../../common/BaseDriver.protractor';
 
 export interface DividerDriver extends BaseDriver {
   exists: () => Promise<boolean>;
