+++ packages/wix-ui-core/src/components/deprecated/Divider/index.ts
@@ -1,0 +1,1 @@
+export {Divider, DividerProps} from './Divider';
